// src/common/enums/application-status.enum.ts
export enum ApplicationStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  UNDER_REVIEW = 'UNDER_REVIEW',
  INTERVIEW_SCHEDULED = 'INTERVIEW_SCHEDULED',
  ADMITTED = 'ADMITTED',
  CONDITIONALLY_ADMITTED = 'CONDITIONALLY_ADMITTED',
  WAITLISTED = 'WAITLISTED',
  REJECTED = 'REJECTED',
  WITHDRAWN = 'WITHDRAWN',
}